+++ source/app/service-providers/assets/types.css-provider.d.ts
@@ -1,7 +1,5 @@
 interface CssProvider {
   on: (event, callback) => void
   off: (event, callback) => void
-  get: () => string
-  set: (newContent: string) => boolean
   getPath: () => string
 }