+++ source/app/service-providers/css-provider.ts
@@ -13,7 +13,7 @@
  */
 
 import path from 'path'
-import fs from 'fs'
+import { promises as fs } from 'fs'
 import { app, ipcMain } from 'electron'
 import EventEmitter from 'events'
 
@@ -29,33 +29,32 @@ export default class CssProvider extends EventEmitter {
 
     // Check for the existence of the custom CSS file. If it is not existent,
     // create an empty one.
-    try {
-      fs.lstatSync(this._filePath)
-    } catch (e) {
-      // Create an empty file with a nice initial comment in it.
-      fs.writeFileSync(this._filePath, '/* Enter your custom CSS here */\n\n')
-    }
+    fs.lstat(this._filePath)
+      .catch(e => {
+        // Create an empty file with a nice initial comment in it.
+        fs.writeFile(this._filePath, '/* Enter your custom CSS here */\n\n', { encoding: 'utf8' })
+          .catch(e => {
+            global.log.error(`[CSS Provider] Could not create Custom CSS file: ${e.message as string}`, e)
+          })
+      })
 
     // Inject the global provider functions
     global.css = {
       on: (event, callback) => { this.on(event, callback) },
       off: (event, callback) => { this.off(event, callback) },
-      get: () => { return this.get() },
-      getPath: () => { return this.getPath() },
-      set: (newContent) => { return this.set(newContent) }
+      getPath: () => { return this.getPath() }
     }
 
     // Send the Custom CSS Path to whomever requires it
-    ipcMain.on('css-provider', (event, message) => {
-      const { command } = message
+    ipcMain.handle('css-provider', async (event, payload) => {
+      const { command } = payload
       if (command === 'get-custom-css-path') {
-        event.sender.send('css-provider', {
-          command: 'get-custom-css-path',
-          payload: this._filePath
-        })
+        return this._filePath
+      } else if (command === 'get-custom-css') {
+        return await this.get()
       } else if (command === 'set-custom-css') {
-        const { payload } = message
-        this.set(payload)
+        const { css } = payload
+        return await this.set(css)
       }
     })
   }
@@ -73,8 +72,8 @@ export default class CssProvider extends EventEmitter {
    * Retrieves the content of the custom CSS file
    * @return {string} The custom CSS
    */
-  get (): string {
-    let file = fs.readFileSync(this._filePath, 'utf8')
+  async get (): Promise<string> {
+    const file = await fs.readFile(this._filePath, { encoding: 'utf8' })
     return file
   }
 
@@ -89,9 +88,9 @@ export default class CssProvider extends EventEmitter {
    * @param {string} newContent The new contents
    * @return {boolean} Whether or not the call succeeded.
    */
-  set (newContent: string): boolean {
+  async set (newContent: string): Promise<boolean> {
     try {
-      fs.writeFileSync(this._filePath, newContent)
+      await fs.writeFile(this._filePath, newContent, { encoding: 'utf8' })
       this.emit('update', this._filePath)
       broadcastIpcMessage('css-provider', {
         command: 'get-custom-css-path',
@@ -99,6 +98,7 @@ export default class CssProvider extends EventEmitter {
       })
       return true
     } catch (e) {
+      global.log.error(`[CSS Provider] Could not set custom css: ${e.message as string}`, e)
       return false
     }
   }