+++ source/renderer/zettlr-rendereripc.js
@@ -442,10 +442,6 @@ class ZettlrRendererIPC {
         this.send('get-pdf-preferences')
         break
 
-      case 'open-custom-css':
-        this._app.getBody().displayCustomCss()
-        break
-
       case 'preferences':
         this._app.getBody().displayPreferences(cnt)
         break